rename shared_sock to sharedsock
