+++ sharedsock/filter_nolinux.go
@@ -1,6 +1,6 @@
 //go:build !linux
 
-package shared_sock
+package sharedsock
 
 // NewSTUNFilter is a noop method just because we do not support BPF filters on other platforms than Linux
 func NewSTUNFilter() BPFFilter {
