+++ sharedsock/sock_linux_test.go
@@ -1,4 +1,4 @@
-package shared_sock
+package sharedsock
 
 import (
 	"context"
